refactored + statistics accumulate universities for each profile.
